+++ ks44team04/src/main/java/ks44team04/admin/controller/user/mapper/UserAddressMapper.java
@@ -1,4 +1,4 @@
-package ks44team04.user.mapper;
+package ks44team04.admin.controller.user.mapper;
 
 import ks44team04.dto.AddressList;
 import org.apache.ibatis.annotations.Mapper;
