+++ ks44team04/src/main/java/ks44team04/admin/controller/user/service/UserAddressService.java
@@ -1,7 +1,7 @@
-package ks44team04.user.service;
+package ks44team04.admin.controller.user.service;
 
+import ks44team04.admin.controller.user.mapper.UserAddressMapper;
 import ks44team04.dto.AddressList;
-import ks44team04.user.mapper.UserAddressMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
