--- ks44team04/src/main/java/ks44team04/user/controller/UserMainController.java
@@ -1,14 +1,0 @@
-package ks44team04.user.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class UserMainController {
-
-    @GetMapping("/user")
-    public String userMain() {
-        return "user/main";
-    }
-
-}
